endpoint GET /users, to obtain existing users
