+++ app/serializers/users.js
@@ -1,6 +1,14 @@
 exports.serializeUser = user => {
-  const { firstName, lastName, ...restUser } = user;
+  const { firstName, lastName, createdAt, updatedAt, ...restUser } = user;
   delete restUser.password;
 
-  return { first_name: firstName, last_name: lastName, ...restUser };
+  return {
+    first_name: firstName,
+    last_name: lastName,
+    created_at: createdAt,
+    updated_at: updatedAt,
+    ...restUser
+  };
 };
+
+exports.serializeUsers = users => users.map(exports.serializeUser);
