+++ test/data/users.js
@@ -20,6 +20,17 @@ exports.user = {
   lastName: 'Newton'
 };
 
+exports.users = [
+  exports.user,
+  { firstName: 'Elon', lastName: 'Musk', email: 'elon.musk@wolox.com', password: 'password' },
+  { firstName: 'Alan', lastName: 'Turin', email: 'alan.turin@wolox.com', password: 'password' }
+];
+
+exports.expectedUsersInfo = exports.users.map(user => {
+  const { firstName, lastName, ...restUser } = user;
+  return { ..._.omit(restUser, 'password'), first_name: firstName, last_name: lastName };
+});
+
 exports.signInRequestBody = {
   email: exports.user.email,
   password: exports.user.password
